Restructure routes into routers
